Move configuring contexts with a dict out
